+++ Runtime/AndroidWidget.cs
@@ -1,7 +1,11 @@
 // Copyright (c) Jerry Lee. All rights reserved. Licensed under the MIT License.
 // See LICENSE in the project root for license information.
 
+// ReSharper disable RedundantUsingDirective
+
 using UnityEngine;
+
+// ReSharper disable RedundantJumpStatement
 
 namespace UniSharper
 {