+++ Runtime/FileExtensions.cs
@@ -1,0 +1,56 @@
+// Copyright (c) Jerry Lee. All rights reserved. Licensed under the MIT License. See LICENSE in the
+// project root for license information.
+
+namespace UniSharper
+{
+    /// <summary>
+    /// The collection of file extension constant.
+    /// </summary>
+    public sealed class FileExtensions
+    {
+        /// <summary>
+        /// The extension of Unity scene file.
+        /// </summary>
+        public const string UnitySceneFile = ".unity";
+        
+        /// <summary>
+        /// The extension of Unity binary file.
+        /// </summary>
+        public const string UnityBinaryFile = ".bytes";
+
+        /// <summary>
+        /// The extension of Unity prefab file.
+        /// </summary>
+        public const string UnityPrefabFile = ".prefab";
+
+        /// <summary>
+        /// The extension of JPG image file.
+        /// </summary>
+        public const string JpgImageFile = ".jpg";
+
+        /// <summary>
+        /// The extension of JPEG image file.
+        /// </summary>
+        public const string JpegImageFile = ".jpeg";
+
+        /// <summary>
+        /// The extension of PNG image file.
+        /// </summary>
+        public const string PngImageFile = ".png";
+
+        /// <summary>
+        /// The extension of old format excel file.
+        /// </summary>
+        public const string ExcelFile = ".xls";
+
+        /// <summary>
+        /// The extension of new format excel file.
+        /// </summary>
+        public const string ExcelXFile = ".xlsx";
+
+        /// <summary>
+        /// The extension of C Sharp script file.
+        /// </summary>
+        public const string CSharpScriptFile = ".cs";
+    }
+}