+++ CHANGELOG.md
@@ -6,11 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [1.17.0] - 2023-12-20
+
+### Added
+
+- Adds class **FileExtensions** that include common file extension constants.
+
+
+
 ## [1.16.0] - 2023-11-24
 
 ### Added
 
-- Adds class **SearchPattern** that include common files search patterns.
+- Adds class **SearchPatterns** that include common file search patterns.